EPMLABSBRN-101 Add availability on exercises retrieved on GET /exercises
