--- src/main/java/com/epam/brn/service/LevelCalculationService.java
@@ -1,7 +1,0 @@
-package com.epam.brn.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class LevelCalculationService {
-}
