+++ src/main/kotlin/com/epam/brn/repo/ExerciseRepository.kt
@@ -1,8 +1,10 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.Exercise
-import org.springframework.data.repository.CrudRepository
+import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ExerciseRepository : CrudRepository<Exercise, Long>
+interface ExerciseRepository : JpaRepository<Exercise, Long> {
+    fun findExercisesBySeriesId(seriesId: Long): List<Exercise>
+}
