+++ src/main/resources/application-dev.properties
@@ -4,3 +4,7 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/${POSTGRES_DB:brn}
 spring.jpa.hibernate.ddl-auto=create
 
 logging.level.org.springframework=debug
+
+#Do not run
+cron.expression.input.data.upload= 0 0 3 29 2 ?
+spring.batch.job.enabled=false
