+++ src/main/resources/application-prod.properties
@@ -2,3 +2,6 @@ spring.datasource.username=${POSTGRES_USER}
 spring.datasource.password=${POSTGRES_PASSWORD}
 spring.datasource.url=jdbc:postgresql://10.66.216.143:5432/${POSTGRES_DB}
 spring.jpa.hibernate.ddl-auto=update
+
+cron.expression.input.data.upload= 0 0/30 * * * ?
+spring.batch.job.enabled=false
