+++ src/main/resources/application.properties
@@ -14,8 +14,6 @@ logging.file.max-history=15
 # Solves issue - https://stackoverflow.com/questions/49669628/spring-boot-application-error-with-connect-postgresql-database?noredirect=1&lq=1
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 
-cron.expression.input.data.upload= 0 0/30 * * * ?
-spring.batch.job.enabled=false
 spring.batch.initialize-schema=always
 spring.flyway.enabled=false
